api-core/pyproject.toml
[project]
name = "api-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Russell Tan",email = "sellnat77@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strawberry-graphql[cli] (>=0.288.2,<0.289.0)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "uvicorn (>=0.40.0,<0.41.0)"
]

[tool.poetry]
packages = [{include = "api_core", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.12.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
